# dejoiy-extract/wp-content/plugins/cursor-plugin--main/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: DEJOIY AI Control Bridge API
  version: 1.0.0
  description: Enterprise AI operating layer for WordPress — REST API for Cursor, Claude, OpenAI Agents, and MCP clients.
  contact:
    name: DEJOIY
    url: https://dejoiy.tech
servers:
  - url: /wp-json/dejoiy-ai/v1
    description: WordPress REST API namespace
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Dejoiy-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /files/tree:
    get:
      summary: Directory tree
      tags: [Files]
      parameters:
        - name: zone
          in: query
          schema: { type: string, enum: [wp-content, themes, plugins, uploads, mu-plugins] }
        - name: path
          in: query
          schema: { type: string }
        - name: depth
          in: query
          schema: { type: integer, default: 3 }
      responses:
        '200':
          description: File tree
  /files/read:
    get:
      summary: Read file
      tags: [Files]
      parameters:
        - { name: zone, in: query, required: true, schema: { type: string } }
        - { name: path, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: File content
  /files/write:
    post:
      summary: Write file
      tags: [Files]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                zone: { type: string }
                path: { type: string }
                content: { type: string }
      responses:
        '200': { description: Success }
        '202': { description: Approval required }
...
  /mcp/execute:
    post:
      summary: Execute MCP tool
      tags: [MCP]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tool: { type: string }
                arguments: { type: object }
...